Rename SubscriptionDataProvider to FeatureListProvider
#4670
