+++ src/subscription/PriceUtils.ts
@@ -4,7 +4,7 @@ import {lang} from "../misc/LanguageViewModel"
 import {neverNull} from "@tutao/tutanota-utils"
 import type {AccountingInfo, Booking, PriceData, PriceItemData} from "../api/entities/sys/TypeRefs.js"
 import {createUpgradePriceServiceData, Customer, CustomerInfo, UpgradePriceServiceReturn} from "../api/entities/sys/TypeRefs.js"
-import {SubscriptionConfig, SubscriptionPlanPrices, SubscriptionType, UpgradePriceType, WebsitePlanPrices} from "./SubscriptionDataProvider"
+import {SubscriptionConfig, SubscriptionPlanPrices, SubscriptionType, UpgradePriceType, WebsitePlanPrices} from "./FeatureListProvider"
 import {locator} from "../api/main/MainLocator"
 import {UpgradePriceService} from "../api/entities/sys/Services"
 import {getTotalAliases, getTotalStorageCapacity, hasAllFeaturesInPlan, isBusinessFeatureActive, isSharingActive, isWhitelabelActive} from "./SubscriptionUtils"
