+++ test/tests/subscription/PriceUtilsTest.ts
@@ -2,7 +2,7 @@ import o from "ospec"
 import {formatMonthlyPrice, formatPrice, getPricesAndConfigProvider, PriceAndConfigProvider} from "../../../src/subscription/PriceUtils.js"
 import {createPlanPrices} from "../../../src/api/entities/sys/TypeRefs.js"
 import {clone} from "@tutao/tutanota-utils"
-import {SubscriptionType, UpgradePriceType} from "../../../src/subscription/SubscriptionDataProvider"
+import {SubscriptionType, UpgradePriceType} from "../../../src/subscription/FeatureListProvider"
 import {matchers, object, when} from "testdouble"
 import {IServiceExecutor} from "../../../src/api/common/ServiceRequest"
 import {UpgradePriceService} from "../../../src/api/entities/sys/Services"
